+++ agent/src/main/java/com/wandern/agent/AgentController.java
@@ -10,10 +10,6 @@ import org.springframework.web.bind.annotation.*;
 import java.util.Collection;
 import java.util.Map;
 
-/**
- * Controller responsible for handling service registration and metrics collection.
- * Provides endpoints to register services and periodically collects metrics from registered services.
- */
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/agent")
@@ -22,6 +18,12 @@ public class AgentController {
     private final AgentService agentService;
     private final ServiceHealthRegistry serviceHealthRegistry;
 
+    /**
+     * Регистрирует сервис в агенте и мастер-сервисе.
+     *
+     * @param serviceInfoDTO информация о регистрируемом сервисе.
+     * @return Ответ с сообщением о успешной регистрации.
+     */
     @PostMapping("/register")
     public ResponseEntity<String> registerService(@RequestBody ServiceInfoDTO serviceInfoDTO) {
         agentService.registerServiceInAgent(serviceInfoDTO);
@@ -29,18 +31,35 @@ public class AgentController {
         return ResponseEntity.ok("Service registered successfully in agent and master.");
     }
 
+    /**
+     * Возвращает список всех зарегистрированных сервисов.
+     *
+     * @return Ответ с коллекцией зарегистрированных сервисов.
+     */
     @GetMapping("/services")
     public ResponseEntity<Collection<ServiceInfoDTO>> getAllRegisteredServices() {
         Collection<ServiceInfoDTO> allServices = agentService.getAllServices();
         return ResponseEntity.ok(allServices);
     }
 
+    /**
+     * Обновляет информацию о состоянии здоровья указанного сервиса.
+     *
+     * @param serviceHealthInfo информация о состоянии здоровья сервиса.
+     * @return Ответ без содержимого.
+     */
+    @Deprecated // крч пока хз
     @PostMapping("/update-service-health")
     public ResponseEntity<Void> updateServiceHealth(@RequestBody ServiceHealthInfo serviceHealthInfo) {
         serviceHealthRegistry.updateServiceHealth(serviceHealthInfo.serviceName(), serviceHealthInfo.healthStatus());
         return ResponseEntity.ok().build();
     }
 
+    /**
+     * Возвращает мапу состояния здоровья всех зарегистрированных сервисов.
+     *
+     * @return Ответ с картой состояния здоровья сервисов.
+     */
     @GetMapping("/services-health")
     public ResponseEntity<Map<String, ServiceHealthInfo>> getServicesHealth() {
         return ResponseEntity.ok(serviceHealthRegistry.getAllServiceHealthInfo());
