+++ agent/src/main/java/com/wandern/agent/ServiceRegistry.java
@@ -11,8 +11,8 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 /**
- * A registry for managing and storing service information.
- * Provides methods to register, retrieve, and remove services.
+ * Класс, отвечающий за регистрацию сервисов.
+ * Хранит информацию о всех зарегистрированных сервисах и предоставляет функциональность для их регистрации.
  */
 @Component
 @RequiredArgsConstructor
@@ -21,11 +21,15 @@ public class ServiceRegistry {
     private final ConcurrentMap<String, ServiceInfoDTO> registry = new ConcurrentHashMap<>();
     private final HealthCheckAgent healthCheckAgent;
 
+    /**
+     * Регистрирует сервис в реестре и инициирует его отслеживание через HealthCheckAgent.
+     *
+     * @param serviceInfoDTO информация о регистрируемом сервисе.
+     */
     public void registerService(ServiceInfoDTO serviceInfoDTO) {
         registry.put(serviceInfoDTO.deploymentId(), serviceInfoDTO);
-        healthCheckAgent.registerService(serviceInfoDTO); // для отслеживания
+        healthCheckAgent.registerService(serviceInfoDTO); // для hc
     }
-
 
     public Optional<ServiceInfoDTO> getService(String deploymentId) {
         return Optional.ofNullable(registry.get(deploymentId));
