+++ agent/src/main/java/com/wandern/agent/health/ServiceHealthInfo.java
@@ -2,6 +2,9 @@ package com.wandern.agent.health;
 
 import com.wandern.starter.health.HealthStatus;
 
+/**
+ * Класс для хранения информации о состоянии здоровья сервиса.
+ */
 public record ServiceHealthInfo (
     String serviceName,
     HealthStatus healthStatus
