--- untitled/src/main/resources/application.yml
@@ -1,4 +1,0 @@
-server:
-  servlet:
-    context-path: /test
-  port: 8099
